+++ generate_package.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+cd src
+zip -r ../firefox-addon-venezuela-no-more.zip *
+cd ..