+++ CHANGELOG.md
@@ -3,3 +3,27 @@
 # 0.1
 
 - Initial release.
+
+
+# 0.2
+
+- Replacements 'Nicolas Maduro' / 'Chiquito de la Calzada'
+- Replacements 'N. Maduro' / 'C. de la Calzada'
+- Replacements 'Maduro' / 'Chiquito'
+- Replacements 'chavismo' / 'chiquitismo'
+
+
+# 0.3
+
+- Replacements 'Nicolás Maduro' / 'Chiquito de la Calzada'
+
+
+# 0.4
+
+- Now works in pages with async updates.
+- Replacements 'Juan Guaidó' / 'Lucas Grijander'
+
+
+# 0.5
+
+- Fix browser action icon.